fix(convex): rename intake-summary module (hyphen breaks Convex deploy)
Convex module paths cannot contain hyphens; convex/documents/intake-summary.ts
(added in #2805) blocked every Convex deploy with InvalidConfig. Renamed to
intakeSummary.ts and updated the two importers.
